requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.5
    # via
    #   genflowly-lambda-utils
    #   python-gemini-api
aiosignal==1.3.1
    # via
    #   aiohttp
    #   genflowly-lambda-utils
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via genflowly-lambda-utils
attrs==23.2.0
    # via
    #   aiohttp
    #   genflowly-lambda-utils
bardapi==1.0.0
    # via genflowly-lambda-utils
blinker==1.8.2
    # via
    #   -r requirements.in
    #   flask
boto3==1.34.144
    # via genflowly-lambda-utils
botocore==1.34.144
    # via
    #   boto3
    #   genflowly-lambda-utils
    #   s3transfer
browser-cookie3==0.19.1
    # via python-gemini-api
build==1.2.1
    # via
    #   -r requirements.in
    #   pip-tools
cachecontrol==0.14.0
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
cachetools==5.3.3
    # via
    #   genflowly-lambda-utils
    #   google-auth
certifi==2024.7.4
    # via
    #   genflowly-lambda-utils
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   genflowly-lambda-utils
charset-normalizer==3.3.2
    # via
    #   genflowly-lambda-utils
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   flask
    #   pip-tools
colorama==0.4.6
    # via bardapi
cryptography==42.0.8
    # via
    #   genflowly-lambda-utils
    #   pyjwt
distro==1.9.0
    # via openai
firebase-admin==6.5.0
    # via genflowly-lambda-utils
flask==3.0.3
    # via
    #   -r requirements.in
    #   prometheus-flask-exporter
    #   pytest-flask
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
    #   genflowly-lambda-utils
genflowly-lambda-utils==0.0.10
    # via -r requirements.in
google-api-core[grpc]==2.19.1
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.137.0
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
google-auth==2.32.0
    # via
    #   genflowly-lambda-utils
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via
    #   genflowly-lambda-utils
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
google-cloud-core==2.4.1
    # via
    #   genflowly-lambda-utils
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.16.1
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
google-cloud-storage==2.17.0
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
google-crc32c==1.5.0
    # via
    #   genflowly-lambda-utils
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via
    #   genflowly-lambda-utils
    #   google-cloud-storage
googleapis-common-protos==1.63.2
    # via
    #   genflowly-lambda-utils
    #   google-api-core
    #   grpcio-status
grpcio==1.64.1
    # via
    #   genflowly-lambda-utils
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.2
    # via
    #   genflowly-lambda-utils
    #   google-api-core
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   genflowly-lambda-utils
    #   google-api-python-client
    #   google-auth-httplib2
httpx[http2]==0.27.0
    # via
    #   bardapi
    #   openai
    #   python-gemini-api
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   genflowly-lambda-utils
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   flask
    #   genflowly-lambda-utils
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
    #   genflowly-lambda-utils
loguru==0.7.2
    # via python-gemini-api
lz4==4.3.3
    # via browser-cookie3
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   genflowly-lambda-utils
    #   jinja2
    #   werkzeug
msgpack==1.0.8
    # via
    #   cachecontrol
    #   genflowly-lambda-utils
multidict==6.0.5
    # via
    #   aiohttp
    #   genflowly-lambda-utils
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.35.13
    # via genflowly-lambda-utils
packaging==24.1
    # via
    #   -r requirements.in
    #   build
    #   pytest
pip-tools==7.4.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
prometheus-client==0.20.0
    # via
    #   -r requirements.in
    #   prometheus-flask-exporter
prometheus-flask-exporter==0.23.1
    # via -r requirements.in
proto-plus==1.24.0
    # via
    #   genflowly-lambda-utils
    #   google-api-core
    #   google-cloud-firestore
protobuf==4.25.3
    # via
    #   genflowly-lambda-utils
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   genflowly-lambda-utils
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   genflowly-lambda-utils
    #   google-auth
pycparser==2.22
    # via
    #   cffi
    #   genflowly-lambda-utils
pycryptodomex==3.20.0
    # via browser-cookie3
pydantic==2.8.2
    # via
    #   openai
    #   python-gemini-api
pydantic-core==2.20.1
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   firebase-admin
    #   genflowly-lambda-utils
pyparsing==3.1.2
    # via
    #   genflowly-lambda-utils
    #   httplib2
pyproject-hooks==1.1.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
pytest==8.3.1
    # via
    #   -r requirements.in
    #   pytest-flask
pytest-flask==1.3.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   genflowly-lambda-utils
python-dotenv==1.0.1
    # via genflowly-lambda-utils
python-gemini-api==2.4.12
    # via bardapi
pyyaml==6.0.1
    # via
    #   genflowly-lambda-utils
    #   responses
requests==2.32.3
    # via
    #   bardapi
    #   cachecontrol
    #   genflowly-lambda-utils
    #   google-api-core
    #   google-cloud-storage
    #   python-gemini-api
    #   requests-oauthlib
    #   responses
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
responses==0.25.3
    # via genflowly-lambda-utils
rsa==4.9
    # via
    #   genflowly-lambda-utils
    #   google-auth
s3transfer==0.10.2
    # via
    #   boto3
    #   genflowly-lambda-utils
six==1.16.0
    # via
    #   genflowly-lambda-utils
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.4
    # via
    #   genflowly-lambda-utils
    #   openai
types-pyyaml==6.0.12.20240311
    # via genflowly-lambda-utils
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via
    #   genflowly-lambda-utils
    #   google-api-python-client
urllib3==2.2.2
    # via
    #   botocore
    #   genflowly-lambda-utils
    #   requests
    #   responses
werkzeug==3.0.3
    # via
    #   -r requirements.in
    #   flask
    #   genflowly-lambda-utils
    #   pytest-flask
wheel==0.43.0
    # via pip-tools
xmltodict==0.13.0
    # via genflowly-lambda-utils
yarl==1.9.4
    # via
    #   aiohttp
    #   genflowly-lambda-utils

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
